[ignore] provide a package to append docker image config in goreleaser file
